fix(app): use correct type for NSArray.count return value
- Change count type from usize to u64 (NSUInteger)
- NSArray.count returns NSUInteger in Objective-C
- objc2 requires exact type matching for message sends
- Fixes runtime panic: 'expected type code q, found Q'

This fixes the crash on startup in overlay.rs:202 when calling
NSArray.count through objc2 message send.

+++ autocorrect-app/src-tauri/src/lib.rs
@@ -273,7 +273,7 @@ pub fn run() {
                         let mut max_x = f64::MIN;
                         let mut max_y = f64::MIN;
                         let screens: Id = msg_send![AnyClass::get("NSScreen").expect("NSScreen not found"), screens];
-                        let count: usize = msg_send![screens, count];
+                        let count: u64 = msg_send![screens, count];
                         for idx in 0..count {
                             let screen: Id = msg_send![screens, objectAtIndex: idx];
                             let frame: CGRect = msg_send![screen, frame];
